Move test runner to src for reuse
