+++ src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./misc";
 export * from "./srcmap";
 export * from "./immutable_map";
+export * from "./test_runner";
