+++ test/utils/index.ts
@@ -1,2 +1,1 @@
 export * from "./pp";
-export * from "./test_runner";
